+++ graphql_compiler/tests/integration_tests/clients/neo4j_graph.py
@@ -3,7 +3,6 @@ from typing import Callable
 
 from neo4j import GraphDatabase
 
-
 NEO4J_SERVER = "localhost"
 NEO4J_PORT = 7687
 NEO4J_USER = "neo4j"
