+++ graphql_compiler/tests/integration_tests/clients/orientdb_schema.sql
