+++ graphql_compiler/tests/integration_tests/clients/redisgraph_graph.py
@@ -4,7 +4,6 @@ from typing import Callable
 import redis
 from redisgraph import Graph
 
-
 REDISGRAPH_SERVER = "localhost"
 REDISGRAPH_PORT = 6379
