+++ graphql_compiler/tests/integration_tests/data/orientdb_snapshot_data.sql
