+++ graphql_compiler/tests/integration_tests/helpers.py
@@ -8,7 +8,7 @@ import six
 from sqlalchemy.engine.base import Engine
 
 from graphql_compiler.schema.schema_info import SQLAlchemySchemaInfo
-from graphql_compiler.tests.test_data_tools.neo4j_graph import Neo4jClient
+from graphql_compiler.tests.integration_tests.clients.neo4j_graph import Neo4jClient
 
 from ... import graphql_to_match, graphql_to_redisgraph_cypher, graphql_to_sql
 from ...compiler import compile_graphql_to_cypher
