--- graphql_compiler/tests/snapshot_tests/snapshots/__init__.py
@@ -1,1 +1,0 @@
-# Copyright 2018-present Kensho Technologies, LLC.
