--- graphql_compiler/tests/test_backend.py
@@ -1,10 +1,0 @@
-# Copyright 2018-present Kensho Technologies, LLC.
-
-POSTGRES = u"postgresql"
-MYSQL = u"mysql"
-MARIADB = u"mariadb"
-MSSQL = u"mssql"
-SQLITE = u"sqlite"
-ORIENTDB = u"orientdb"
-NEO4J = u"neo4j"
-REDISGRAPH = u"redisgraph"
