--- graphql_compiler/tests/test_data_tools/__init__.py
@@ -1,1 +1,0 @@
-# Copyright 2018-present Kensho Technologies, LLC.
